fastiron

Module constants

Hardcoded constants used by the simulation

Constants are sorted in submodules according to their nature. Are also included aliases, as well as the custom trait used to introduce a generic floating point type in the code.

The below text is taken directly from a Quicksilver source file:

The values of all physical constants are taken from 2006 CODATA which is located here.

The units of physical quantities used by the code are:

Quantity	Unit

Mass	gram (g)

Length	centimeter (cm)

Time	second (s)

Energy	million electron-volts (MeV) : of a particle

Energy	erg (g cm^2/s^2): in some background calculation

Temperature	thousand electron-volts (keV)

Modules

sim

Simulation-related constants

Traits

CustomFloat

Custom trait for floating point number

CustomReferenceFloat

Associated reference value used for compute approximation.

Type Aliases

Tuple3

Custom alias for readability.

Tuple4

Custom alias for readability.