fastiron::data

Module energy_spectrum

Discrete energy spectrum

This module contains the structure used to store and build a discrete energy spectrum from the problem’s population.

Structs

EnergySpectrum

Structure used to represent the energy spectrum of the problem, i.e. the distribution of particles among energy levels.