Module fastiron::data::nuclear_data

Nuclear-related data and modelling

This module contains code used to store, model and compute nuclear data and quantities.

Structs

NuclearData

Top level structure used to handle all things related to nuclear data.

NuclearDataReaction

Lowest-level structure to represent a reaction.

NuclearDataSpecies

Structure used to hold a list of reactions.

Polynomial

Structure used to represent a polynomial function.

Enums

ReactionType

Enum used to represent a reaction type.

Type Aliases

NuclearDataIsotope

Structure used to store reactions for a given isotope.