fastiron::data::nuclear_data

Type Alias NuclearDataIsotope

pub type NuclearDataIsotope<T> = Vec<NuclearDataSpecies<T>>;

Structure used to store reactions for a given isotope.

At the current time there is only one type of particle in the simulation, meaning that there is only one species per isotope.

Aliased Type

pub struct NuclearDataIsotope<T> { /* private fields */ }