Module fastiron::geometry::facets

Facet modelling & related computation

This module contains the code used for the modelling and computations related to the facets of the mesh.

Structs

MCFacetAdjacency

Structure for adjacent facet representation.

MCFacetAdjacencyCell

Structure encompassing all adjacent facet to a cell.

MCGeneralPlane

Structure representing a plane of equation a*x + b*y + c*z + d = 0

MCNearestFacet

Structure used to represent the nearest facet to a particle, holding relevant data for computation.

SubfacetAdjacency

Sub-structure for adjacent facet representation.

Enums

MCSubfacetAdjacencyEvent

Enum used to categorize the event a particle undergo when reaching a given facet.

Type Aliases

MCFacetGeometryCell

List of planes associated with the outward-facing facets of a cell.