fastiron::geometry::global_fcc_grid

Struct GlobalFccGrid

pub struct GlobalFccGrid<T: CustomFloat> {
    pub nx: usize,
    pub ny: usize,
    pub nz: usize,
    pub lx: T,
    pub ly: T,
    pub lz: T,
    pub dx: T,
    pub dy: T,
    pub dz: T,
}

Structure representing the spatial grid of the problem.

Fields
nx: usize

Number of cells along the x-axis.

ny: usize

Number of cells along the y-axis.

nz: usize

Number of cells along the z-axis.

lx: T

Size of the problem along the x-axis (cm).

ly: T

Size of the problem along the y-axis (cm).

lz: T

Size of the problem along the z-axis (cm).

dx: T

Size of a mesh cell along the x-axis (cm).

dy: T

Size of a mesh cell along the y-axis (cm).

dz: T

Size of a mesh cell along the z-axis (cm).

Implementations

impl<T: CustomFloat> GlobalFccGrid<T>

pub fn new(nx: usize, ny: usize, nz: usize, lx: T, ly: T, lz: T) -> Self

Constructor.

pub fn which_cell(&self, r: &MCVector<T>) -> usize

Returns the index of the cell the coordinates belong to.

pub fn cell_center(&self, idx_cell: usize) -> MCVector<T>

Returns the coordinate of the center of the given cell.

pub fn cell_idx_to_tuple(&self, idx_cell: usize) -> Tuple3

Converts a cell index to a coordinate tuple.

pub fn cell_tuple_to_idx(&self, tuple_cell: &Tuple3) -> usize

Converts a cell coordinate tuple to an index.

pub fn node_tuple_to_idx(&self, tt: &Tuple4) -> usize

Converts a node coordinate tuple to an index.

pub fn node_idx_to_tuple(&self, idx: usize) -> Tuple4

Converts a node index to a coordinate tuple.

pub fn get_node_gids(&self, cell_gid: usize) -> [usize; 14]

Returns the global identifiers of the nodes of the specified cell.

pub fn get_face_nbr_gids(&self, cell_gid: usize) -> [usize; 6]

Returns the global identifiers of the faces of the specified cell.

pub fn node_coord_from_idx(&self, idx: usize) -> MCVector<T>

Returns a node’s coordinate from its index.

pub fn node_coord_from_tuple(&self, tt: &Tuple4) -> MCVector<T>

Returns a node’s coordinate from its tuple.

pub fn snap_turtle(&self, tt: (i32, i32, i32)) -> Tuple3

Adjust the tuple value according to bounds.