Module fastiron::init

Initialization code for the problem

Functions

check_cross_sections

Prints cross-section data of the problem.

consistency_check

Check the consistency of the domain list passed as argument.

init_mcdata

Creates a MonteCarloData object using the specified parameters.

init_mcunits

Creates & initialize the correct number of Monte-Carlo units for simulation.

init_particle_containers

Creates the correct number of particle containers for simulation.

init_results

Initialize the structure used to keep track & process results of the simulation.