fastiron::init

Function init_particle_containers

pub fn init_particle_containers<T: CustomFloat>(
    params: &Parameters<T>,
    proc_info: &MCProcessorInfo,
) -> Vec<ParticleContainer<T>>

Creates the correct number of particle containers for simulation.

The correct number is determined according to simulation parameters & execution policy.