fastiron

Module parameters

Parameters-related code

The module holds all parameters related structure used during the simulation. Input reading is located in the crate::utils module, while parsing is done here.

Structs

CrossSectionParameters

Structure used to describe a cross-section, i.e. a probability density representation.

GeometryParameters

Structure used to describe a geometry, i.e. a physical space of a certain shape and certain material.

MaterialParameters

Struct used to describe a material, i.e. its name and relevant physical properties.

Parameters

Structure holding all the problem’s parameters.

SimulationParameters

Structure holding all simulation parameters.

Enums

BenchType

Enum used to run additional tests according to the input benchmark

ParameterError

Enum used to categorize inconsistencies within parameters

Shape

Enum used to describe a geometry’s shape

Type Aliases

Block

Alias for a <String, String> [FxHashMap]. See here for detailed structure of input blocks.