fastiron::parameters

Struct SimulationParameters

pub struct SimulationParameters<T: CustomFloat> {

    pub input_file: String,
    pub energy_spectrum: String,
    pub cross_sections_out: String,
    pub boundary_condition: String,
    pub load_balance: bool,
    pub csv: bool,
    pub debug_threads: bool,
    pub n_particles: u64,
    pub chunk_size: u64,
    pub n_rayon_threads: u64,
    pub bind_threads: bool,
    pub n_units: u64,
    pub n_steps: usize,
    pub nx: usize,
    pub ny: usize,
    pub nz: usize,
    pub seed: u64,
    pub dt: T,
    pub lx: T,
    pub ly: T,
    pub lz: T,
    pub e_min: T,
    pub e_max: T,
    pub n_groups: usize,
    pub low_weight_cutoff: T,
    pub coral_benchmark: BenchType,

}

Structure holding all simulation parameters.

In the program’s execution flow, it is first initialized using the CLI arguments, then optionally updated with a specified input file.

Fields
input_file: String

Path to the input file, it can be relative or absolute.

energy_spectrum: String

Name of the output file the energy spectrum may be saved to.

cross_sections_out: String

Name of the output file the cross-sections may be saved to.

boundary_condition: String

Boundary conditions of the problem. Mesh is initialized according to this value.

load_balance: bool

Switch to enable or disable load balancing during execution.

csv: bool

Switch to enable writing of tallies & timers data in csv files.

debug_threads: bool

Switch used to print thread-debugging information. Currently unused.

n_particles: u64

Target number of particle for the simulation. Population will be controlled according to this value.

chunk_size: u64

Size of the chunks used when executing in parallel.

n_rayon_threads: u64

Number of threads that should be used to run the simulation.

bind_threads: bool

Switch used to bind rayon threads to physical cores.

n_units: u64

Number of units that should be used to run the simulation.

n_steps: usize

Number of steps simulated by the program.

nx: usize

Number of cells along the x-axis.

ny: usize

Number of cells along the y-axis.

nz: usize

Number of cells along the z-axis.

seed: u64

Random number seed for the PRNG used by the simulation.

dt: T

Value of the time step in seconds.

lx: T

Size of the simulation along the x-axis.

ly: T

Size of the simulation along the y-axis.

lz: T

Size of the simulation along the z-axis.

e_min: T

Energy value of the lowest energy group.

e_max: T

Energy value of the highest energy group.

n_groups: usize

Number of energy groups to build a spectrum.

low_weight_cutoff: T

Low statistical weight cutoff used for population control.

coral_benchmark: BenchType

Benchmark type of the input problem. See BenchType for more information.

Implementations

impl<T: CustomFloat> SimulationParameters<T>

pub fn from_cli(cli: &Cli) -> Self

Initialize a SimulationParameters object using a Cli object created from the arguments supplied via command line.

use clap::Parser;
use fastiron::utils::input::Cli;
use fastiron::parameters::SimulationParameters;

let cli = Cli::parse_from("./fastiron -i somefile -c -l".split(' '));
let simulation_params = SimulationParameters::<f64>::from_cli(&cli);
// compare the structures...
assert_eq!(cli.input_file.unwrap(), simulation_params.input_file);
assert!(simulation_params.load_balance);