fastiron::particles::mc_particle

Struct MCParticle

pub struct MCParticle<T: CustomFloat> {

    pub coordinate: MCVector<T>,
    pub direction: MCVector<T>,
    pub kinetic_energy: T,
    pub energy_group: usize,
    pub weight: T,
    pub species: Species,
    pub time_to_census: T,
    pub age: T,
    pub total_cross_section: T,
    pub mean_free_path: T,
    pub segment_path_length: T,
    pub num_mean_free_paths: T,
    pub num_segments: T,
    pub random_number_seed: u64,
    pub identifier: u64,
    pub last_event: MCTallyEvent,
    pub domain: usize,
    pub cell: usize,
    pub facet: usize,
    pub normal_dot: T,

}

Structure used to hold all data of a particle.

This is mostly used for computations during the tracking section.

Fields
coordinate: MCVector<T>

Current position.

direction: MCVector<T>

Direction of the particle as a normalized (x, y, z) vector.

kinetic_energy: T

Kinetic energy.

energy_group: usize

Current energy group the particle belong to.

weight: T

Weight.

species: Species

Species of the particle.

time_to_census: T

Time remaining before this particle hit census.

age: T

Age.

total_cross_section: T

Cache-ing the current total cross-section.

mean_free_path: T

Distance to a collision.

segment_path_length: T

Distance this particle travels in a segment.

num_mean_free_paths: T

Number of mean free paths to a collision.

num_segments: T

Number of segments the particle travelled.

random_number_seed: u64

Random number seed for the rng for this particle.

identifier: u64

Unique ID used to identify and track individual particles.

last_event: MCTallyEvent

Last event this particle underwent.

domain: usize

Current domain in the spatial grid.

cell: usize

Current cell in the current domain.

facet: usize

Nearest facet.

normal_dot: T

Normal dot product value kept when crossing a facet.

Implementations

impl<T: CustomFloat> MCParticle<T>

pub fn move_particle_along_segment(&mut self)

Update the particleโs field to model its movement along its direction, for the segment length.

pub fn update_trajectory(&mut self, energy: T, angle: T)

Update the particleโs trajectory with new energy & angle.

pub fn get_speed(&self) -> T

Computes the particle speed from its energy. Note that this computation should be species-specific.

pub fn get_current_xs(&mut self) -> T

Return the starting cross-section for reaction sampling.

pub fn sample_collision( &mut self, reaction: &NuclearDataReaction<T>, material_mass: T, extra: &mut ParticleCollection<T>, ) -> usize

Uses a PRNG to sample new energy & angle after a reaction.

Since reaction type is specified when the method is called, we assume that the result will be treated correctly by the calling code.

pub fn sample_num_mfp(&mut self)

Sample the number of mean free paths to a collision.

pub fn sample_isotropic(&mut self)

Sample a random direction for the particle to face.

pub fn rotate_direction( &mut self, sine_theta: T, cosine_theta: T, sine_phi: T, cosine_phi: T, )

Rotates a 3D vector that is defined by the angles Theta and Phi in a local coordinate frame about a polar angle and azimuthal angle described by the direction cosine. Hence, caller passes in sin_Theta and cos_Theta referenced from the local z-axis and sin_Phi and cos_Phi referenced from the local x-axis to describe the vector V to be rotated. The direction cosine describes global theta and phi angles that the vector V is to be rotated about. cos_theta_zero/sin_theta_zero and cos_phi_zero/sin_phi_zero model the initial position while the arguments of the method characterize the rotation. See this for explanation on the formula.

pub fn under_populated_split( &mut self, split_rr_factor: T, ) -> impl Iterator<Item = MCParticle<T>> + '_

Returns an iterator over particles created from a split of the original one (caller). This is used in the population control algorithm.

pub fn over_populated_rr(&mut self, split_rr_factor: T) -> bool

Play russian-roulette with the particle, returning true if the particle survives, false otherwise. This function is meant to be used along the Vec::retain_mut() method.

pub fn low_weight_rr( &mut self, relative_weight_cutoff: T, source_particle_weight: T, ) -> bool

Play russian-roulette with particle of low statistical weight, returning true if the particle survives, false otherwise. This function is meant to be used along the Vec::retain_mut() method.