Module fastiron::simulation

Computation & simulation-related code

This module contains the core functions that make up the simulation loop.

Modules

collision_event

Event-specific code for particles colliding

cycle_tracking

Core of the particle tracking algorithm used by the simulation

macro_cross_section

Code for cross-section computation a.k.a. probability density computation

mc_facet_crossing_event

Event-specific code for particles crossing a cell’s facet

mc_segment_outcome

Code for segment outcome computation, i.e. whichever event the particle undergoes next

mct

Code for spatial computations of the simulation

population_control

Code regulating the number of particles in the simulation