fastiron::simulation

Module collision_event

Event-specific code for particles colliding

This module contains code that process particles undergoing a collision from beginning to end. Note that collision refers to reaction with the particle’s environment, not in-between particles.

Functions

collision_event

Transforms a given particle according to an internally drawn type of collision.