Function fastiron::simulation::collision_event::collision_event

pub fn collision_event<T: CustomFloat>(
    mcdata: &MonteCarloData<T>,
    balance: &mut Balance,
    mat_gid: usize,
    cell_nb_density: T,
    particle: &mut MCParticle<T>,
    extra: &mut ParticleCollection<T>,
) -> bool

Transforms a given particle according to an internally drawn type of collision.

The function calls method from super::macro_cross_section module to pick the reaction the particle will undergo (See ReactionType). The particle is then updated and the collision tallied. Finally, particles are created / invalidated accordingly to the picked reaction:

Absorption reaction: the particle is invalidated.

Fission reaction: offspring particles are created from the colliding one.

Scattering reaction: no additional modifications occur.