Function fastiron::simulation::cycle_tracking::cycle_tracking_guts

pub fn cycle_tracking_guts<T: CustomFloat>(
    mcdata: &MonteCarloData<T>,
    mcunit: &MonteCarloUnit<T>,
    particle: &mut MCParticle<T>,
    balance: &mut Balance,
    extra: &mut ParticleCollection<T>,
)

Main steps of the CycleTracking section.

The particle at the specified index is loaded, tracked and updated accordingly. Depending on the outcome of the tracking, it is either set as processed or invalidated.