fastiron::simulation::mc_facet_crossing_event

Function facet_crossing_event

pub fn facet_crossing_event<T: CustomFloat>(
    particle: &mut MCParticle<T>,
    facet_adjacency: &SubfacetAdjacency,
)

Computes and transform accordingly a MCParticle object crossing a facet.

This function updates a particle’s locational data according to one of the four defined adjacency events (MCSubfacetAdjacencyEvent). Note that in a sequential or a memory-shared parallelism context, there are no off-processor transit as the mesh is not divided for management.