fastiron::simulation::mc_segment_outcome

Function outcome

pub fn outcome<T: CustomFloat>(
    mcdata: &MonteCarloData<T>,
    mcunit: &MonteCarloUnit<T>,
    particle: &mut MCParticle<T>,
) -> MCSegmentOutcome

Computes the outcome of the current segment for a given particle.

Three outcomes are possible for a given particle: census, facet crossing or collision. The distances are computed as follows:

Census: The distance is simply equal to the speed of the particle multiplied by the time left until census.

Facet crossing: The distance is computed in a similar way by the function nearest_facet().

Collision: The distance is computed using probabilities.