fastiron::simulation

Module mct

Code for spatial computations of the simulation

This module contains function used to compute and manipulate data related to a particle’s coordinate and direction in the problem.

Functions

cell_position_3dg

Returns a coordinate that represents the “center” of the cell.

generate_coordinate_3dg

Generates a random coordinate inside a polyhedral cell.

nearest_facet

Computes which facet of the specified cell is nearest to the specified coordinates.

reflect_particle

Reflects a particle off a reflection-type boundary.