Function fastiron::simulation::population_control::source_now

pub fn source_now<T: CustomFloat>(
    mcdata: &MonteCarloData<T>,
    mcunit: &mut MonteCarloUnit<T>,
    container: &mut ParticleContainer<T>
)

Simulates the sources according to the problem’s parameters.

This function spawns particle is source regions. Each time this function is called (once per cycle), 10% of the target number of particles are spawned. Where they are spawned depends on both deterministic factors and randomness.

Each cell is given a weight, computed according to source rate of its material, volume and time step: The weight is directly proportional to the number of particle that should be spawned in the cell, during a time interval equal to the time step. Given this weight, and the weight of a fresh particle at sourcing, we can compute the number of particle the cell should spawn to get the overall desired number.

The amount of particle spawned is a consequence of the source particle weight computation as the value is adjusted according to the problem’s total weight.