Struct fastiron::utils::input::Cli

pub struct Cli {

    pub input_file: Option<String>,
    pub energy_spectrum: Option<String>,
    pub cross_sections_out: Option<String>,
    pub dt: Option<f32>,
    pub load_balance: bool,
    pub csv: bool,
    pub debug_threads: bool,
    pub single_precision: bool,
    pub lx: Option<f32>,
    pub ly: Option<f32>,
    pub lz: Option<f32>,
    pub n_particles: Option<u64>,
    pub chunk_size: Option<u64>,
    pub n_rayon_threads: Option<u64>,
    pub n_units: Option<u64>,
    pub n_steps: Option<usize>,
    pub nx: Option<usize>,
    pub ny: Option<usize>,
    pub nz: Option<usize>,
    pub seed: Option<u64>,

}

Fastiron, a Rust port of the Quicksilver proxy-app

Fields
input_file: Option<String>

name of input file

energy_spectrum: Option<String>

name of energy spectrum output file

cross_sections_out: Option<String>

name of cross-section output file

dt: Option<f32>

time step in seconds

load_balance: bool

enable load balancing if present

csv: bool

write tallies & timer data into csv files if present

debug_threads: bool

enable thread debugging if present

single_precision: bool

enable single-precision float type usage if present

lx: Option<f32>

x-size of simulation in cm

ly: Option<f32>

y-size of simulation in cm

lz: Option<f32>

z-size of simulation in cm

n_particles: Option<u64>

total number of particules

chunk_size: Option<u64>

size of the chunks when executing in parallel – if absent or set to 0, use dynamic chunk size

n_rayon_threads: Option<u64>

number of rayon threads that should be used to run the simulation – set to 0 for rayon’s default config

n_units: Option<u64>

number of units that should be used to run the simulation

n_steps: Option<usize>

number of steps simulated

nx: Option<usize>

number of mesh elements along x

ny: Option<usize>

number of mesh elements along y

nz: Option<usize>

number of mesh elements along z

seed: Option<u64>

random number seed