Module fastiron::utils::mc_fast_timer

Code used for built-in timers & execution speed gauging

Structs

MCFastTimer

Structure used to represent a single timer.

MCFastTimerContainer

Structure used as a container for the 6 timers used through the simulation for performance testing.

Enums

Section

Enum used to identify sections and their corresponding timers.

Functions

get_last_cycle

Returns the duration of the last cycle of the specified timer.

start

Start the specified timer.

stop

Stop the specified timer and record internally the duration since start.