fastiron::constants

Trait CustomReferenceFloat

pub trait CustomReferenceFloat: Float + FromPrimitive {
    const HUGE_FLOAT: Self;
    const SMALL_FLOAT: Self;
    const TINY_FLOAT: Self;
    const PI: Self;
}

Associated reference value used for compute approximation.

This is the only trait that should be manually implemented for custom types. The CustomFloat trait is then automatically implemented via a blanket implementation.

Required Associated Constants

const HUGE_FLOAT: Self

Threshold upper-value for decimal number.

const SMALL_FLOAT: Self

Threshold low-ish-value for decimal number.

const TINY_FLOAT: Self

Threshold lower-value for decimal number.

const PI: Self

Pi value.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.