fastiron::data

Module tallies

Code for simulation-related statistics

This module contains all code used to record and count events happening during the simulation. The cyclic summary is printed using the recorded data.

Note that this module isn’t used to compute time-related data, this is done in the utils::mc_fast_timer module.

Structs

Balance

Structure used to keep track of the number of event in the simulation.

FluenceDomain

Domain-sorted fluence-data-holding sub-structure.

ScalarFluxDomain

Domain-sorted scalar-flux-data-holding sub-structure.

Tallies

Super-structure holding all recorded data besides time statistics.

Enums

MCTallyEvent

Enum representing a tally event.

TalliedEvent

Constants

N_TALLIED_EVENT