Module fastiron::geometry

Mesh & modelling-related structures$

This module contains code related to the mesh structure used to simulate the problem. This includes the mesh itself as well as objects used for navigating it. The caracteristics of the mesh are defined as constants.

The used mesh is made of cells (hexahedron), each divided in 12 sub-cells (tetrahedron). TODO: insert an image?

Modules

facets

Facet modelling & related computation

global_fcc_grid

Code used to build and navigate the basis of the mesh

grid_assignment_object

Code used to locate vectors in the grid

mc_cell_state

Contains code to represent a cell’s current state.

mc_domain

Code used for the domain separation model

mc_location

Code used to asociate a location to a particle

mesh_partition

Code used to build and model the mesh of the problem

Constants

CORNER_OFFSET

Offsets of the intersection points of a cell.

FACE_OFFSET

Offsets of the faces of a cell.

N_FACES

Number of faces defining a cell.

N_FACETS_OUT

Number of facets of a cell facing outward i.e. constituting a border with another cell.

N_POINTS_INTERSEC

Number of points defining a cell.

N_POINTS_PER_FACET

Number of points per tetrahedron facet.