Function fastiron::init::check_cross_sections

pub fn check_cross_sections<T: CustomFloat>(mcdata: &MonteCarloData<T>)

Prints cross-section data of the problem.

The data is simply presented in a table sorted by energy levels. Here is a snippet of the first ten lines of a possible output:

group |           energy |  sourceMaterial_absorb |  sourceMaterial_fission |  sourceMaterial_scatter
    0 |   0.000000001054 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    1 |   0.000000001168 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    2 |   0.000000001294 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    3 |   0.000000001434 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    4 |   0.000000001589 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    5 |   0.000000001761 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    6 |   0.000000001952 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    7 |   0.000000002163 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    8 |   0.000000002397 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945
    9 |   0.000000002656 |        0.0108589141086 |         0.0096890310969 |         0.0794520547945

The energy scale is logarithmic, hence the way it is printed.