Module fastiron::simulation::mc_segment_outcome

Code for segment outcome computation, i.e. whichever event the particle undergoes next

This module contains the function computing which event a given particle will undergo next. This is done by using the notion of distance to an event and finding the minimum distance between all the possible events.
The particle’s position and time to census is also updated, while event-specific modifications are left to event-specific functions. The result is returned using an enum (MCSegmentOutcome)that takes value according to the event.

Structs

DistanceHandler

Structure used to handle all distance related data & comparison.

Enums

MCSegmentOutcome

Enum representing the outcome of the current segment.

Functions

outcome

Computes the outcome of the current segment for a given particle.